style: Refactor mixins.scss for improved layout and styling
Refactor the mixins.scss file to improve the layout and styling of the application. This includes updating the container mixin to have a max-width of 1200px, a width of 90%, and a margin of 0 auto. Additionally, update the row mixin to have a flex-direction of column and a gap of 2em. Finally, update the map mixin to have a width of 100%, a height of 400px, a border-radius of 1em, and a box-shadow of 0 0 .5em rgba(0, 0, 0, 0.9). These changes enhance the overall presentation and responsiveness of the application.